# Use a mutable struct for the state of GridSpaceIdIterator