This reduces the allocations of the iterator.

As you can see on this Flocking benchmark inside the ABMFrameworkComparison repo the speed improvement is nice:

```julia
#before
julia> a = @benchmark run_model(rng, (150, 150), 400, 15.0) setup = (rng = rng_model()) evals=1 samples=n_run seconds=1e6
BenchmarkTools.Trial: 100 samples with 1 evaluation.
 Range (min … max):  37.952 ms … 61.669 ms  ┊ GC (min … max): 5.92% … 8.08%
 Time  (median):     47.160 ms              ┊ GC (median):    7.55%
 Time  (mean ± σ):   47.599 ms ±  4.318 ms  ┊ GC (mean ± σ):  7.46% ± 1.12%

                ▆   ▂▆  █▄ █  ▄ ▂                              
  ▄▄▁▁▄▁▁▁█▁▆▆▄██▆█▆██▆▄██▆█▆██▄██▄▆█▄▆▁▆▁▁▆█▁▄▁▆▁▁▁▁▁▁▁▁▁▁▁▄ ▄
  38 ms           Histogram: frequency by time          61 ms <

 Memory estimate: 98.61 MiB, allocs estimate: 2151950.

#after
julia> a = @benchmark run_model(rng, (150, 150), 400, 15.0) setup = (rng = rng_model()) evals=1 samples=n_run seconds=1e6
BenchmarkTools.Trial: 100 samples with 1 evaluation.
 Range (min … max):  32.827 ms … 52.396 ms  ┊ GC (min … max): 3.73% … 2.89%
 Time  (median):     40.189 ms              ┊ GC (median):    6.13%
 Time  (mean ± σ):   40.487 ms ±  3.500 ms  ┊ GC (mean ± σ):  6.12% ± 1.13%

                   █   ▄▆▁▆   ▁                                
  ▄▄▄▄▁▄▁▄▄▄▁▆▇▇▆▇▆█▇▇▇████▆▇▇█▆▆▄▆▄▇▁▄▇▄▆▄▄▁▄▁▁▁▄▁▁▁▁▁▁▁▁▁▁▄ ▄
  32.8 ms         Histogram: frequency by time          52 ms <

 Memory estimate: 67.06 MiB, allocs estimate: 2191950.
```

(using Ref.(....) to retrieve the fields can sometimes have a bad effect on inference, so I also changed that)